langchain/requirements.txt
langchain
ipykernel
jupyter
python-dotenv
langchain_community
pypdf
bs4
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
chromadb
langchain_chroma
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
langchain_huggingface
huggingface_hub